requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
gunicorn==21.2.0
charset-normalizer==3.3.2
